Tests passing! Reviewing the solution found below, which is a guess.
Reviewing the errors, what I got from it is that:

What I suspect is that I'm probably approaching this the wrong way, and that ostream doesn't
have a copy constructor, or really any way to be declared, because even
just declaring via std::ostream os; will break, meaning that I'm
probably supposed to set the pointer of the variable to std::cout,
something like that, but I'm not sure.

To be even more specific: what I understand is that ios_base
is a type unintended for being assigned, and basic_ostream is a type
unintended for being intialized, given the protected/private nature
of those functons.

The only way I know how to declare without some sort of
auto-initialization stuff happening is through memory memes, and that is
probably not the appropriate solution.

Therefore, I decided to just hold a pointer to the stream, which seems
to have worked. Only thing is I need to dereference it every time I
write, which I would have preferred to avoid; oh well!